updates on regulating parser and updater behavior.
